+++ yugioh-20221012/src/main/java/tw/gaas/yugioh/domain/Player.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import tw.gaas.yugioh.domain.card.Cards;
 
 @Builder
 @Data
