+++ yugioh-20221012/src/main/java/tw/gaas/yugioh/domain/card/Card.java
@@ -1,9 +1,9 @@
-package tw.gaas.yugioh.domain;
+package tw.gaas.yugioh.domain.card;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
-import tw.gaas.yugioh.domain.enu.Type;
+import tw.gaas.yugioh.domain.card.enu.Type;
 
 @Data
 @SuperBuilder
