+++ yugioh-20221012/src/main/java/tw/gaas/yugioh/domain/card/Cards.java
@@ -1,4 +1,4 @@
-package tw.gaas.yugioh.domain;
+package tw.gaas.yugioh.domain.card;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
