+++ yugioh-20221012/src/main/java/tw/gaas/yugioh/domain/card/MonsterCard.java
@@ -1,11 +1,11 @@
-package tw.gaas.yugioh.domain;
+package tw.gaas.yugioh.domain.card;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
-import tw.gaas.yugioh.domain.enu.Attribute;
-import tw.gaas.yugioh.domain.enu.MonsterType;
+import tw.gaas.yugioh.domain.card.enu.Attribute;
+import tw.gaas.yugioh.domain.card.enu.MonsterType;
 
 @SuperBuilder
 @Data
