+++ yugioh-20221012/src/main/java/tw/gaas/yugioh/domain/card/SpellCard.java
@@ -1,17 +1,17 @@
-package tw.gaas.yugioh.domain;
+package tw.gaas.yugioh.domain.card;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
-import tw.gaas.yugioh.domain.enu.TrapType;
+import tw.gaas.yugioh.domain.card.enu.SpellType;
 
 @SuperBuilder
 @Data
 @EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor
-public class TrapCard extends Card {
+public class SpellCard extends Card {
 
-    // 陷阱卡類型
-    private TrapType trapType;
+    // 魔法卡類型
+    private SpellType spellType;
 }
