+++ yugioh-20221012/src/main/java/tw/gaas/yugioh/domain/card/enu/Attribute.java
@@ -1,4 +1,4 @@
-package tw.gaas.yugioh.domain.enu;
+package tw.gaas.yugioh.domain.card.enu;
 
 public enum Attribute {
